Add RCP_Member->get_switch_to_url()
If the user switching plugin is active,
this method will return a url to switch to
the user that is the current instance of
RCP_Member.

Intended for use with a wrapper function:
rcp_get_switch_to_url(), which accepts a user
id parameter. The user id passed is the user
for which the switch to url will be generated.