Remove condition with empty( CONSTANT ) check
In PHP empty() does not support checking a constant. It only allows variables and expressions. This will cause PHP to throw an error expecting :: as it is looking for an object with the name TWOCHECKOUT_ADMIN_USER and TWOCHECKOUT_ADMIN_PASSWORD. Because the defined() check is above this empty() check may be redundant. If the user does put an empty string in then using the constant would fail, just as it would if they input an invalid password.