Add Alipay support to Stripe Checkout. #432. (#504)
Adds a setting for enabling Alipay support. Adds a default data-alipay attribute to the [register_form_stripe] shortcode, which gets its default value from the global setting. Unsets the data-alipay shortcode attribute if someone has enabled it but USD is not the currency, because according to their docs it's only available in USD. https://stripe.com/docs/alipay